port: begin implementing audio
mixer taken from sm64-port and modified to fit PD's naudio thing

MP3 playback not supported yet

audio buffering sucks

+++ include/PR/n_libaudio.h
@@ -71,7 +71,7 @@ typedef struct {
     s32                 maxAuxBusses;
     s32                 outputRate;
     s32                 maxOutSamples;
-    s32                 sv_dramout;
+    uintptr_t           sv_dramout;
     s32                 sv_first;
 } N_ALSynth;
